Add TimeSliceBuffer to buffer and sort the Ts of Set<T>
Logic here assumes that all Set<T> in a time slice have the same
start_time and end_time and come in on the queues in time order. All
objects of the Set<T> in the time slice are then sorted by time_start
before being passed to the operator() of algorithms.