Add confgen for TPSets from readout to TPSetSink
Requires branch swtpg-port of readout

This commit adds a two-process confgen with a readout app and a TP
sink app. The readout app takes in a binary raw data file and runs the
software trigger primitive generation on the raw data. The TPs
produced are arranged into TPSets and published. The TP sink app
subscribes to this stream of TPSets and feeds them into a TPSetSink,
which does basic checks on the ordering and content of the TPSets.