First attempt at expressing confgens more concisely
Express confgens more concisely by writing them as a list of modules
and connections, rather than a series of commands and queues. Much
common code is moved into the new `util` module. We can also infer a
lot of things that were previously written explicitly, eg what queues
have to exist (based on the connections between modules) and the
correct order to start and stop modules to allow all data to be
processed.