Add and use modulegraph object
modulegraph adds "endpoints" to the dictionary of modules that
form (part of) an application. These endpoints provide a public name
for other applications to use, without depending on the internal
details of modules and sink/source names inside the modulegraph. This
approach will also allow checking that all required endpoints have
been connected in the system generation stage.